HBX-1663: Move class 'org.hibernate.tool.hbm2x.pojo.ComponentPOJOClass' to 'org.hibernate.tool.internal.export.pojo.ComponentPOJOClass'
