+++ orm/src/main/java/org/hibernate/tool/internal/export/pojo/Cfg2JavaTool.java
@@ -25,7 +25,6 @@ import org.hibernate.mapping.SimpleValue;
 import org.hibernate.mapping.Value;
 import org.hibernate.tool.hbm2x.ExporterException;
 import org.hibernate.tool.hbm2x.MetaAttributeHelper;
-import org.hibernate.tool.hbm2x.pojo.ComponentPOJOClass;
 import org.hibernate.tool.hbm2x.pojo.EntityPOJOClass;
 import org.hibernate.tool.hbm2x.visitor.JavaTypeFromValueVisitor;
 import org.hibernate.tool.internal.export.hbm.Cfg2HbmTool;
